requirements.txt
pandas
numpy
openai
gradio
scipy
python-dotenv
pytest
pytest-mock